Avoid using _POSIX_VERSION on Emscripten
When building with Emscripten, depending on luck, <unistd.h> might or
might not be included before <quick-lint-js/have.h> is processed. This
means that _POSIX_VERSION might or might not be defined. This means that
QLJS_HAVE_MKDTEMP and other macros can have different definitions in
different translation units. Worse, the definition can change with the
CMake version which determines whether precompiled headers are enabled.

Make the definition of QLJS_HAVE_MKDTEMP and other macros predictable on
Emscripten by only checking _POSIX_VERSION if QLJS_HAVE_UNISTD_H is
defined.

For non-Emscripten POSIX targets, this commit should not change
behavior.